requirements.txt
flask
scikit-learn
joblib
tldextract
python-whois
requests
gunicorn
